web/requirements.txt
streamlit==1.51.0
requests==2.32.5
pydantic==2.12.5
python-magic==0.4.27
